services structure refactoring
