+++ src/web/server/loader/application/tasks.py
@@ -3,11 +3,11 @@ import datetime
 from marshmallow import ValidationError
 import graphyte
 
-from src.web.server.loader.sensor_loading import load_sensors
-from src.web.server.loader.weather_loading import parse_weather
+from src.web.server.loader.application.sensor_loading import load_sensors
+from src.web.server.loader.application.weather_loading import parse_weather
 from src.web.server.common.model import Sensors, Weather
-from src.web.server.loader.validation import SensorSchema, WeatherSchema
-from src.web.server.loader.mosecom_loading import load_data, write_processed, write_raw_data
+from src.web.server.loader.application.validation import SensorSchema, WeatherSchema
+from src.web.server.loader.application.mosecom_loading import load_data, write_processed, write_raw_data
 from src.web.config import metrics_host
 
 sensor_schema = SensorSchema()
