+++ src/web/server/ml/application/anomaly.py
@@ -3,7 +3,7 @@ import datetime
 import pickle
 
 from src.web.server.common.model import Anomaly
-from src.web.server.ml.data_loading import get_weather_data, get_sensor_data
+from src.web.server.ml.application.data_loading import get_weather_data, get_sensor_data
 from src.model.anom_clustering import Model
 
 pca_model_file = 'models/anomalies/dim_red.obj'
