+++ tests/web/test_mosecom_loading.py
@@ -1,7 +1,7 @@
 import os
 import numpy as np
 
-from src.web.server.loader.mosecom_loading import load_data, write_raw_data, write_processed, avarege_data
+from src.web.server.loader.application.mosecom_loading import load_data, write_raw_data, write_processed, avarege_data
 from src.web.server.loader.config import mosecom_url
